+++ plugin/othttp/server.go
@@ -1,4 +1,4 @@
-package httptrace
+package othttp
 
 import (
 	"io"
